+++ src/main/kotlin/top/flapypan/blog/entity/Tag.kt
@@ -10,7 +10,7 @@ import jakarta.persistence.*
 class Tag {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var id: Long = 0L
+    var id: Int = 0
 
     @Column(nullable = false, unique = true)
     var name: String = ""
